refactor(evals): keep resolve_pass_threshold in engine/instance, drop scoring shim
Reverts the previous move to model_hub/utils/scoring. resolve_pass_threshold
stays where it originally lived (evaluations/engine/instance) since
user_evaluation and the runner already import from there; composite_execution
now imports from the same place instead of maintaining a scoring-utils shim.

The type / range guard added last commit is preserved inline at the helper's
new home.
